+++ common/src/main/java/de/cech12/brickfurnace/block/BrickFurnaceBlock.java
@@ -3,22 +3,20 @@ package de.cech12.brickfurnace.block;
 import de.cech12.brickfurnace.Constants;
 import de.cech12.brickfurnace.blockentity.AbstractBrickFurnaceBlockEntity;
 import de.cech12.brickfurnace.blockentity.BrickFurnaceBlockEntity;
+import net.minecraft.core.BlockPos;
+import net.minecraft.stats.Stats;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.BaseEntityBlock;
 import net.minecraft.world.level.block.FurnaceBlock;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.MenuProvider;
-import net.minecraft.stats.Stats;
 import net.minecraft.world.level.block.entity.BlockEntity;
-import net.minecraft.core.BlockPos;
-import net.minecraft.world.level.Level;
-
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-
 import net.minecraft.world.level.block.entity.BlockEntityTicker;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public class BrickFurnaceBlock extends FurnaceBlock {
 
@@ -27,13 +25,13 @@ public class BrickFurnaceBlock extends FurnaceBlock {
     }
 
     @Override
-    public BlockEntity newBlockEntity(@Nonnull BlockPos blockPos, @Nonnull BlockState blockState) {
+    public BlockEntity newBlockEntity(@NotNull BlockPos blockPos, @NotNull BlockState blockState) {
         return new BrickFurnaceBlockEntity(blockPos, blockState);
     }
 
     @Override
     @Nullable
-    public <T extends BlockEntity> BlockEntityTicker<T> getTicker(@Nonnull Level level, @Nonnull BlockState state, @Nonnull BlockEntityType<T> entityType) {
+    public <T extends BlockEntity> BlockEntityTicker<T> getTicker(@NotNull Level level, @NotNull BlockState state, @NotNull BlockEntityType<T> entityType) {
         return BaseEntityBlock.createTickerHelper(entityType, (BlockEntityType<AbstractBrickFurnaceBlockEntity>) Constants.BRICK_FURNACE_BLOCK_ENTITY_TYPE.get(), AbstractBrickFurnaceBlockEntity::tick);
     }
 
@@ -42,7 +40,7 @@ public class BrickFurnaceBlock extends FurnaceBlock {
      * inside AbstractFurnaceBlock.
      */
     @Override
-    protected void openContainer(Level worldIn, @Nonnull BlockPos pos, @Nonnull Player player) {
+    protected void openContainer(Level worldIn, @NotNull BlockPos pos, @NotNull Player player) {
         BlockEntity blockEntity = worldIn.getBlockEntity(pos);
         if (blockEntity instanceof BrickFurnaceBlockEntity) {
             player.openMenu((MenuProvider)blockEntity);